requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.9.0
reportlab==4.2.5
Flask-Limiter==3.8.0
